requirements.txt
asgiref==3.5.0
beautifulsoup4==4.9.3
bs4==0.0.1
certifi==2021.5.30
chardet==4.0.0
Django==3.2.12
django-crontab==0.7.1
django-restframework==0.0.1
djangorestframework==3.13.1
idna==2.10
lxml==4.6.3
numpy==1.21.2
Pillow==9.0.1
pytz==2021.3
requests==2.25.1
soupsieve==2.2.1
sqlparse==0.4.2
typing_extensions==4.1.1
urllib3==1.26.6
